DEV: rename API folder
