+++ spinalcordtoolbox/deepseg_sc/cnn_models.py
